Move context definition to a separate file
