+++ packages/domain-search/src/page/context.ts
@@ -1,7 +1,7 @@
-import { createContext, useContext } from 'react';
+import { createContext, useCallback, useContext, useMemo, useState } from 'react';
 import { domainAvailabilityQuery } from '../queries/availability';
 import { domainSuggestionsQuery, freeSuggestionQuery } from '../queries/suggestions';
-import { type DomainSearchContextType } from './types';
+import type { DomainSearchProps, DomainSearchContextType } from './types';
 
 const noop = () => {};
 
@@ -44,4 +44,73 @@ export const useDomainSearch = () => {
 	}
 
 	return context;
+};
+
+export const useDomainSearchContextValue = (
+	props: DomainSearchProps
+): typeof DEFAULT_CONTEXT_VALUE => {
+	const { currentSiteUrl, initialQuery, cart, events, slots, config } = props;
+
+	const [ isFullCartOpen, setIsFullCartOpen ] = useState( false );
+	const [ query, setQuery ] = useState( initialQuery ?? '' );
+
+	const closeFullCart = useCallback( () => {
+		setIsFullCartOpen( false );
+	}, [] );
+
+	const openFullCart = useCallback( () => {
+		setIsFullCartOpen( true );
+	}, [] );
+
+	const normalizedEvents = useMemo( () => {
+		return {
+			...DEFAULT_CONTEXT_VALUE.events,
+			...events,
+		};
+	}, [ events ] );
+
+	const normalizedConfig = useMemo( () => {
+		return {
+			...DEFAULT_CONTEXT_VALUE.config,
+			...config,
+		};
+	}, [ config ] );
+
+	return useMemo( () => {
+		return {
+			events: normalizedEvents,
+			config: normalizedConfig,
+			queries: {
+				domainSuggestions: ( query ) =>
+					domainSuggestionsQuery( query, {
+						quantity: 30,
+						vendor: normalizedConfig.vendor,
+					} ),
+				freeSuggestion: ( query ) => ( {
+					...freeSuggestionQuery( query ),
+					enabled: normalizedConfig.skippable,
+				} ),
+				domainAvailability: domainAvailabilityQuery,
+			},
+			cart,
+			isFullCartOpen,
+			closeFullCart,
+			openFullCart,
+			query,
+			setQuery,
+			slots,
+			currentSiteUrl,
+		};
+	}, [
+		isFullCartOpen,
+		closeFullCart,
+		openFullCart,
+		query,
+		setQuery,
+		cart,
+		normalizedEvents,
+		slots,
+		currentSiteUrl,
+		normalizedConfig,
+	] );
 };
