+++ packages/domain-search/src/page/index.tsx
@@ -1,9 +1,7 @@
 import { QueryClientProvider } from '@tanstack/react-query';
 import clsx from 'clsx';
-import { useCallback, useState, useMemo, useLayoutEffect } from 'react';
-import { domainAvailabilityQuery } from '../queries/availability';
-import { domainSuggestionsQuery, freeSuggestionQuery } from '../queries/suggestions';
-import { DEFAULT_CONTEXT_VALUE, DomainSearchContext } from './context';
+import { useLayoutEffect } from 'react';
+import { DomainSearchContext, useDomainSearchContextValue } from './context';
 import { EmptyPage } from './empty';
 import { fallbackQueryClient } from './fallback-query-client';
 import { ResultsPage } from './results';
@@ -11,74 +9,12 @@ import type { DomainSearchProps } from './types';
 
 import './style.scss';
 
-export const DomainSearch = ( {
-	className,
-	currentSiteUrl,
-	initialQuery,
-	cart,
-	events,
-	slots,
-	queryClient = fallbackQueryClient,
-	config,
-}: DomainSearchProps ) => {
-	const [ isFullCartOpen, setIsFullCartOpen ] = useState( false );
-	const [ query, setQuery ] = useState( initialQuery ?? '' );
+export const DomainSearch = ( props: DomainSearchProps ) => {
+	const contextValue = useDomainSearchContextValue( props );
 
-	const closeFullCart = useCallback( () => {
-		setIsFullCartOpen( false );
-	}, [] );
-
-	const openFullCart = useCallback( () => {
-		setIsFullCartOpen( true );
-	}, [] );
-
-	const contextValue: typeof DEFAULT_CONTEXT_VALUE = useMemo( () => {
-		const normalizedConfig = {
-			...DEFAULT_CONTEXT_VALUE.config,
-			...config,
-		};
-
-		return {
-			events: {
-				...DEFAULT_CONTEXT_VALUE.events,
-				...events,
-			},
-			config: normalizedConfig,
-			queries: {
-				domainSuggestions: ( query ) =>
-					domainSuggestionsQuery( query, {
-						quantity: 30,
-						vendor: normalizedConfig.vendor,
-					} ),
-				freeSuggestion: ( query ) => ( {
-					...freeSuggestionQuery( query ),
-					enabled: normalizedConfig.skippable,
-				} ),
-				domainAvailability: domainAvailabilityQuery,
-			},
-			cart,
-			isFullCartOpen,
-			closeFullCart,
-			openFullCart,
-			query,
-			setQuery,
-			slots,
-			currentSiteUrl,
-		};
-	}, [
-		isFullCartOpen,
-		closeFullCart,
-		openFullCart,
-		query,
-		setQuery,
-		cart,
-		events,
-		slots,
-		currentSiteUrl,
-		config,
-	] );
-
-	const cartItemsLength = cart.items.length;
+	const cartItemsLength = contextValue.cart.items.length;
+	const isFullCartOpen = contextValue.isFullCartOpen;
+	const closeFullCart = contextValue.closeFullCart;
 
 	useLayoutEffect( () => {
 		if ( cartItemsLength === 0 && isFullCartOpen ) {
@@ -87,7 +23,7 @@ export const DomainSearch = ( {
 	}, [ cartItemsLength, isFullCartOpen, closeFullCart ] );
 
 	const getContent = () => {
-		if ( ! query ) {
+		if ( ! contextValue.query ) {
 			return <EmptyPage />;
 		}
 
@@ -95,9 +31,9 @@ export const DomainSearch = ( {
 	};
 
 	return (
-		<QueryClientProvider client={ queryClient }>
+		<QueryClientProvider client={ fallbackQueryClient }>
 			<DomainSearchContext.Provider value={ contextValue }>
-				<div className={ clsx( 'domain-search', className ) }>{ getContent() }</div>
+				<div className={ clsx( 'domain-search', props.className ) }>{ getContent() }</div>
 			</DomainSearchContext.Provider>
 		</QueryClientProvider>
 	);
